cleaning code up
